TyrantFactionChat: Assume permission=member if not found
If a newly-joined member confirms registration while the cache is stale,
we return prematurely and do not let them register. Clearly this is bad
user experience.

Is there any security risk here? Only if an arbitrary user can post to
faction chat without being a member of the faction.

The better solution of course is to replace the cached data with new
data, but even if we do that someone may still confirm before the cache
is refreshed. So we need this patch in either case, and refreshing the
cache is almost an orthogonal improvement.