Auth: Remove explicit known_ids
Instead, dynamically create as needed.
Removing this makes manual management of the YAML file easier and less
error-prone, as deleting a user had to be done by deleting an entry in
nicks and an entry in known_ids.

We can consider keeping it but having only nicks be saved in YAML. This
saves repeated regeneration if !register is used repeatedly but
maintains the increased maintainability.